--- src/component/VideoComponent.vue
@@ -1,20 +1,0 @@
-<template>
-  <div class="video-component">
-    test video component
-  </div>
-</template>
-
-<script lang="ts">
-import { defineComponent } from "vue";
-
-export default defineComponent({
-  name: "Panel"
-});
-</script>
-
-<style scoped>
-.video-component {
-  background: white;
-  color: black;
-}
-</style>
