--- src/component/index.ts
@@ -1,4 +1,0 @@
-import { createApp } from "vue";
-import VideoComponent from "./VideoComponent.vue";
-
-createApp(VideoComponent).mount("#app");
